+++ src/index.js
@@ -3,7 +3,5 @@ import ReactDOM from 'react-dom';
 import Router from './Router';
 import './styles/reset.scss';
 import './styles/common.scss';
-import './pages/yongwoo/Login/Login.scss';
-import './pages/yongwoo/Main/Main.scss';
 
 ReactDOM.render(<Router />, document.getElementById('root'));
