+++ src/pages/yongwoo/Login/Login.js
@@ -1,72 +1,16 @@
 import React from 'react';
-import { Link, useNavigate } from 'react-router-dom';
-
-function Logo() {
-  return <h1 className="logo">Westagram</h1>;
-}
-
-function UserInfo() {
-  return (
-    <>
-      <input
-        className="userName"
-        type="text"
-        placeholder="전화번호, 사용자 이름 또는 이메일"
-      />
-      <input className="userPassword" type="password" placeholder="비밀번호" />
-    </>
-  );
-}
-
-function SignUpWrap() {
-  return (
-    <article className="signUpWrap">
-      <p>
-        계정이 없으신가요?{' '}
-        <Link to="/main-yongwoo">
-          <button href="" className="signUpBlue">
-            가입하기
-          </button>
-        </Link>
-      </p>
-    </article>
-  );
-}
-
-function LoginButton() {
-  const navigate = useNavigate();
-
-  const goToMain = () => {
-    navigate('/main-yongwoo');
-  };
-  return (
-    <button className="loginButton" onClick={goToMain}>
-      로그인
-    </button>
-  );
-}
-
-function StoreWrap() {
-  return (
-    <article className="storeWrap">
-      <img
-        src="https://www.instagram.com/static/images/appstore-install-badges/badge_ios_korean-ko.png/4a5c9d62d51b.png"
-        alt="appStore"
-      />
-      <img
-        src="https://www.instagram.com/static/images/appstore-install-badges/badge_android_korean-ko.png/f155b664a93b.png"
-        alt="googlePlay"
-      />
-    </article>
-  );
-}
+import StoreWrap from './StoreWrap/StoreWrap';
+import LoginButton from './LoginButton/LoginButton';
+import UserInfomation from './UserInfomation/UserInfomation';
+import { Link } from 'react-router-dom';
+import './Login.scss';
 
 function LoginYongwoo() {
   return (
-    <section className="login">
+    <section className="LoginYongwoo">
       <article className="topWrap">
-        <Logo />
-        <UserInfo />
+        <h1 className="logo">Westagram</h1>
+        <UserInfomation />
         <LoginButton />
         <div className="meddleLine">
           <div className="line" />
@@ -78,7 +22,16 @@ function LoginYongwoo() {
         </div>
         <p className="passwordLost">비밀번호를 잊으셨나요?</p>
       </article>
-      <SignUpWrap />
+      <article className="signUpWrap">
+        <p>
+          계정이 없으신가요?{' '}
+          <Link to="/main-yongwoo">
+            <button href="" className="signUpBlue">
+              가입하기
+            </button>
+          </Link>
+        </p>
+      </article>
       <p className="downlodeCopyright">앱을 다운로드하세요.</p>
       <StoreWrap />
     </section>
