+++ src/pages/yongwoo/Login/Login.scss
@@ -1,12 +1,6 @@
 @import url('https://fonts.googleapis.com/css2?family=Lobster&display=swap');
+@import '../../../styles/variables.scss';
 
-$border-gray: 1px solid #dbdbdb;
-
-@mixin flex-center {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
 @mixin flex-column-center {
   display: flex;
   flex-direction: column;
@@ -14,18 +8,18 @@ $border-gray: 1px solid #dbdbdb;
 }
 
 body {
-  @include flex-center;
+  @include flexCenter;
   height: 100vh;
 }
 
-.login {
+.LoginYongwoo {
   @include flex-column-center;
   width: 350px;
 
   .topWrap {
     @include flex-column-center;
     width: 348px;
-    border: $border-gray;
+    border: 1px solid $color-outline;
     background-color: #ffffff;
 
     .logo {
@@ -34,33 +28,6 @@ body {
       font-size: 45px;
     }
 
-    input {
-      width: 266px;
-      padding: 10px 5px;
-      margin-top: 5px;
-      color: #8e8e8e;
-      background-color: #fafafa;
-      border: 1px solid rgba(var(--b6a, 219, 219, 219), 1);
-      border-radius: 1px;
-      font-size: 12px;
-
-      &::placeholder {
-        padding-left: 2px;
-        font-size: 12px;
-      }
-    }
-
-    .loginButton {
-      width: 266px;
-      padding: 8px 0;
-      margin: 15px 0;
-      color: #ffffff;
-      background-color: #b2dffc;
-      border: none;
-      border-radius: 1px;
-      outline: none;
-      font-weight: bold;
-    }
     .meddleLine {
       display: flex;
       align-items: center;
@@ -73,7 +40,7 @@ body {
       .line {
         width: 103px;
         height: 1px;
-        border-bottom: $border-gray;
+        border-bottom: 1px solid $color-outline;
       }
     }
 
@@ -90,11 +57,11 @@ body {
     }
   }
   .signUpWrap {
-    @include flex-center;
+    @include flexCenter;
     width: 348px;
     height: 60px;
     margin-top: 20px;
-    border: $border-gray;
+    border: 1px solid $color-outline;
     background-color: #ffffff;
     font-size: 14px;
 
@@ -111,9 +78,5 @@ body {
   .downlodeCopyright {
     margin: 20px 0;
     font-size: 14px;
-  }
-
-  .storeWrap img {
-    width: 136px;
   }
 }
