+++ src/pages/yongwoo/Login/StoreWrap/StoreWrap.scss
@@ -1,0 +1,6 @@
+.storeWrap {
+  img {
+    width: 136px;
+    margin-left: 5px;
+  }
+}
