+++ src/pages/yongwoo/Login/UserInfomation/UserInfomation.scss
@@ -1,0 +1,22 @@
+@import '../../../../styles/variables.scss';
+
+.userInfomation {
+  display: flex;
+  flex-direction: column;
+
+  input {
+    width: 266px;
+    padding: 10px 5px;
+    margin-top: 5px;
+    color: #8e8e8e;
+    background-color: #fafafa;
+    border: 1px solid rgba(var(--b6a, 219, 219, 219), 1);
+    border-radius: $border-radius;
+    font-size: 12px;
+
+    &::placeholder {
+      padding-left: 2px;
+      font-size: 12px;
+    }
+  }
+}
