+++ src/pages/yongwoo/Main/CommentContainer/CommentContainer.scss
@@ -1,0 +1,29 @@
+@import '../../../../styles/variables.scss';
+
+.commentContainer {
+  display: flex;
+  width: 100%;
+  height: 50px;
+  margin-top: 10px;
+  border-top: 1px solid $color-outline;
+
+  .commentWrap {
+    @include flexBetween;
+    width: 570px;
+    margin: 0 auto;
+
+    input {
+      width: 500px;
+      border: none;
+      background: none;
+      outline: none;
+    }
+
+    button {
+      background: none;
+      border: none;
+      cursor: pointer;
+      color: #bfe0fd;
+    }
+  }
+}
